pyproject.toml
[project]
name = "contacts-manager"
version = "0.1.0"
description = "An API to store and manage your contacts"
authors = [
    {name = "Oleksandr Romashko",email = "alex.rmshk@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy (>=2.0.43)",
    "fastapi (>=0.116.1,<0.117.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic[email] (>=2.11.9,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.urls]
Homepage = "https://github.com/oleksandr-romashko/goit-pythonweb-hw-10"
Issues = "https://github.com/oleksandr-romashko/goit-pythonweb-hw-10/issues"
License = "https://github.com/oleksandr-romashko/goit-pythonweb-hw-10/blob/main/LICENSE"